Completed

+++ delete_comment.php
@@ -1,11 +1,9 @@
 <?php
-
+require "database.php";
 session_start();
 
-require "database.php";
-
-if (isset($_GET["comment_id"])) {
-
+if (isset($_GET["comment_id"]) and isset($_GET["post_id"])) {
+	// Delete the comment from the table
 	$stmt = $mysqli->prepare("DELETE FROM comments WHERE comment_id = ?");
 
 	if(!$stmt){
@@ -14,14 +12,14 @@ if (isset($_GET["comment_id"])) {
 	}
 
 	$stmt->bind_param('s', $_GET["comment_id"]);
-
 	$stmt->execute();
-	 
 	$stmt->close();
 
 	header("Location: individual_story.php?post_id=" . $_GET['post_id']);
-
 	exit;
 }
-
+else {
+	header("Location: main.php");
+	exit;
+}
 ?>
